extract class component
